analytics-engine/requirements.txt
psycopg2-binary==2.9.9
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
redis==5.0.1
numpy==1.26.4
